+++ lib/FieldLocation/FieldLocationFinal.js
@@ -4,30 +4,25 @@ import { FormattedMessage } from 'react-intl';
 
 import { LocationLookup } from '@folio/stripes/smart-components';
 
-import {
-  FieldSelection,
-  FieldSelectionFinal,
-} from '../FieldSelection';
+import { FieldSelectionFinal } from '../FieldSelection';
 import { validateRequired } from '../utils';
 import { fieldSelectOptionsShape } from '../shapes';
 
-const FieldLocation = ({
+const FieldLocationFinal = ({
+  fieldComponent,
   isDisabled,
-  isReduxField,
   labelId,
   locationLookupLabel,
   locationsForSelect,
   name,
   required,
   selectLocationFromPlugin,
-  validate,
 }) => {
-  const Component = isReduxField ? FieldSelection : FieldSelectionFinal;
-  const validateValue = validate || (required && validateRequired);
+  const FieldComponent = fieldComponent || FieldSelectionFinal;
 
   return (
     <div>
-      <Component
+      <FieldComponent
         dataOptions={locationsForSelect}
         disabled={isDisabled}
         fullWidth
@@ -36,7 +31,7 @@ const FieldLocation = ({
         marginBottom0
         name={name}
         required={required}
-        validate={validateValue}
+        validate={required ? validateRequired : undefined}
       />
       {!isDisabled && (
         <LocationLookup
@@ -48,19 +43,15 @@ const FieldLocation = ({
   );
 };
 
-FieldLocation.propTypes = {
+FieldLocationFinal.propTypes = {
+  fieldComponent: PropTypes.elementType,
   isDisabled: PropTypes.bool.isRequired,
-  isReduxField: PropTypes.bool,
   labelId: PropTypes.string,
   locationLookupLabel: PropTypes.node,
   locationsForSelect: fieldSelectOptionsShape.isRequired,
   name: PropTypes.string,
   required: PropTypes.bool.isRequired,
   selectLocationFromPlugin: PropTypes.func.isRequired,
-  validate: PropTypes.oneOfType([
-    PropTypes.func,
-    PropTypes.arrayOf(PropTypes.func),
-  ]),
 };
 
-export default FieldLocation;
+export default FieldLocationFinal;
