+++ lib/FieldLocation/FieldLocationFinal.test.js
@@ -6,7 +6,7 @@ import { Form } from 'react-final-form';
 
 import '../../test/jest/__mock__';
 
-import FieldLocation from './FieldLocation';
+import FieldLocationFinal from './FieldLocationFinal';
 
 const locationOptions = [
   { value: '001', label: 'Location #1' },
@@ -16,12 +16,12 @@ const locationOptions = [
 
 const fieldLocationLabel = 'Location';
 
-const renderFieldLocation = (labelId, locationsForSelect, selectLocationFromPlugin = noop) => (render(
+const renderFieldLocationFinal = (labelId, locationsForSelect, selectLocationFromPlugin = noop) => (render(
   <IntlProvider locale="en">
     <Form
       onSubmit={noop}
       render={() => (
-        <FieldLocation
+        <FieldLocationFinal
           labelId={labelId}
           name="locationId"
           locationsForSelect={locationsForSelect}
@@ -33,17 +33,17 @@ const renderFieldLocation = (labelId, locationsForSelect, selectLocationFromPlug
   </IntlProvider>,
 ));
 
-describe('FieldLocation component', () => {
+describe('FieldLocationFinal component', () => {
   afterEach(cleanup);
 
   it('should display passed label', () => {
-    const { getByText } = renderFieldLocation('Location', locationOptions);
+    const { getByText } = renderFieldLocationFinal('Location', locationOptions);
 
     expect(getByText(fieldLocationLabel)).toBeDefined();
   });
 
   it('should render all passed options', async () => {
-    const { findAllByText } = renderFieldLocation(undefined, locationOptions);
+    const { findAllByText } = renderFieldLocationFinal(undefined, locationOptions);
 
     const renderedLocationOptions = await findAllByText(/Location #[0-9]/);
