+++ lib/FieldLocation/FieldLocationFinalContainer.test.js
@@ -6,7 +6,7 @@ import { Form } from 'react-final-form';
 
 import '../../test/jest/__mock__';
 
-import FieldLocationContainer from './FieldLocationContainer';
+import FieldLocationFinalContainer from './FieldLocationFinalContainer';
 
 const locationsList = [
   { id: '001', name: 'Location #1' },
@@ -16,7 +16,7 @@ const locationsList = [
 const locationsIds = ['001', '002'];
 const fieldLocationLabel = 'Location';
 
-const renderFieldLocationContainer = (
+const renderFieldLocationFinalContainer = (
   labelId,
   prepopulatedLocationsIds,
   locationsForDict,
@@ -26,7 +26,7 @@ const renderFieldLocationContainer = (
     <Form
       onSubmit={noop}
       render={() => (
-        <FieldLocationContainer
+        <FieldLocationFinalContainer
           labelId={labelId}
           name="locationId"
           prepopulatedLocationsIds={prepopulatedLocationsIds}
@@ -38,17 +38,17 @@ const renderFieldLocationContainer = (
   </IntlProvider>,
 ));
 
-describe('FieldLocationContainer component', () => {
+describe('FieldLocationFinalContainer component', () => {
   afterEach(cleanup);
 
   it('should display passed label', () => {
-    const { getByText } = renderFieldLocationContainer('Location');
+    const { getByText } = renderFieldLocationFinalContainer('Location');
 
     expect(getByText(fieldLocationLabel)).toBeDefined();
   });
 
   it('should render options based on passed locationIds', async () => {
-    const { findAllByText } = renderFieldLocationContainer('Location', locationsIds, locationsList);
+    const { findAllByText } = renderFieldLocationFinalContainer('Location', locationsIds, locationsList);
 
     const renderedLocationOptions = await findAllByText(/Location #[0-9]/);
