+++ lib/FieldLocation/index.js
@@ -1,1 +1,2 @@
+export { default as FieldLocationFinal } from './FieldLocationFinalContainer';
 export { default as FieldLocation } from './FieldLocationContainer';
